+++ public/index.php
@@ -5,16 +5,16 @@
  * Simple password manager written in PHP with Bootstrap and PDO database connections
  *
  *  File name: index.php
- *  Last Modified: 3.01.23 г., 11:31 ч.
+ *  Last Modified: 3.01.23 г., 12:59 ч.
  *
- * @link          https://blacktiehost.com
- * @since         1.0.0
- * @version       2.2.0
- * @author        Milen Karaganski <milen@blacktiehost.com>
+ *  @link          https://blacktiehost.com
+ *  @since         1.0.0
+ *  @version       2.3.0
+ *  @author        Milen Karaganski <milen@blacktiehost.com>
  *
- * @license       GPL-3.0+
- * @license       http://www.gnu.org/licenses/gpl-3.0.txt
- * @copyright     Copyright (c)  2020 - 2022 blacktiehost.com
+ *  @license       GPL-3.0+
+ *  @license       http://www.gnu.org/licenses/gpl-3.0.txt
+ *  @copyright     Copyright (c)  2020 - 2022 blacktiehost.com
  *
  */
 
@@ -36,7 +36,7 @@ try {
     include_once('../includes/main.inc.php');
 } catch (Exception $e) {
     $error = $e->getMessage();
-    pm_syslog('Cannot load file vendor/autoload.php with error ' . $error, LOG_ERR);
+    pm_syslog('Cannot load file includes/main.inc.php with error ' . $error, LOG_ERR);
     print 'File "includes/main.inc.php!"not found';
     die();
 }